requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0
redis==6.1.0
pydantic==2.11.4
starlette==0.46.2
async-timeout==5.0.1